% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greekletters.R
\name{print_greeks}
\alias{print_greeks}
\title{Print all Greek letters and mathematical symbols names}
\usage{
print_greeks()
}
\value{
Print all Greek letters and mathematical symbols names.
}
\description{
Function to Print all Greek letters and mathematical symbols names.
}
\details{
It is recommended that the font size of the R console be increased for
better visualization of the symbols,
as some of the symbols are quite small.

The subscripts and supersripts are restricted to numbers 0 to 9.

Symbols printed as a "rectangle" don't work in RGui, but these symbols work normally in RStudio.
}
\examples{
\donttest{
#all Greek letters and mathematical symbols names
print_greeks()
}

}
\author{
Kévin Allan Sales Rodrigues.
}
